Make --version report the active config file
This is useful for troubleshooting.
