+++ lib/ansible/constants.py
@@ -17,67 +17,10 @@
 
 import os
 import pwd
-import sys
-import ConfigParser
 from string import ascii_letters, digits
 
-# copied from utils, avoid circular reference fun :)
-def mk_boolean(value):
-    if value is None:
-        return False
-    val = str(value)
-    if val.lower() in [ "true", "t", "y", "1", "yes" ]:
-        return True
-    else:
-        return False
+from .config import get_config, load_config_file
 
-def get_config(p, section, key, env_var, default, boolean=False, integer=False, floating=False, islist=False):
-    ''' return a configuration variable with casting '''
-    value = _get_config(p, section, key, env_var, default)
-    if boolean:
-        return mk_boolean(value)
-    if value and integer:
-        return int(value)
-    if value and floating:
-        return float(value)
-    if value and islist:
-        return [x.strip() for x in value.split(',')]
-    return value
-
-def _get_config(p, section, key, env_var, default):
-    ''' helper function for get_config '''
-    if env_var is not None:
-        value = os.environ.get(env_var, None)
-        if value is not None:
-            return value
-    if p is not None:
-        try:
-            return p.get(section, key, raw=True)
-        except:
-            return default
-    return default
-
-def load_config_file():
-    ''' Load Config File order(first found is used): ENV, CWD, HOME, /etc/ansible '''
-
-    p = ConfigParser.ConfigParser()
-
-    path0 = os.getenv("ANSIBLE_CONFIG", None)
-    if path0 is not None:
-        path0 = os.path.expanduser(path0)
-    path1 = os.getcwd() + "/ansible.cfg"
-    path2 = os.path.expanduser("~/.ansible.cfg")
-    path3 = "/etc/ansible/ansible.cfg"
-
-    for path in [path0, path1, path2, path3]:
-        if path is not None and os.path.exists(path):
-            try:
-                p.read(path)
-            except ConfigParser.Error as e:
-                print "Error reading config file: \n%s" % e
-                sys.exit(1)
-            return p
-    return None
 
 def shell_expand_path(path):
     ''' shell_expand_path is needed as os.path.expanduser does not work
